backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
openai==1.54.0
httpx<0.28.0
python-dotenv==1.0.1
pydantic==2.9.2
